Put string generators in a module
