+++ test/enigma_test.rb
@@ -6,7 +6,6 @@ class EnigmaTest < Minitest::Test
   end
 
   def test_it_exists
-    binding.pry
     assert_instance_of Enigma, @enigma
   end
